Update for use with Guard >=2.9
Guard deprecated reliance on Guard::Guard over a year ago and from 2.8.x
it was displaying deprecation warnings. As of 2.9.x Guard::Guard defined in
guard/guard.rb has been removed breaking plugins that didn't upgrade.

In addition, the new version of Guard breaks many RSpecs because internals
are not initialised if we directly instantiate a plugin. For a discussion
about this issue see https://github.com/guard/guard/issues/693

This commit makes the following changes:-
 - Foodcritic now extends Guard::Plugin instead of Guard::Guard.
 - Initialize method signature now has 1 param (options hash) instead of 2
 - Foodcritic depends on guard-compat to stub out dependencies for RSPEC
 - Updated Specs for deprecated matchers
 - Updated Specs for changed method signature